fixed an id-mismatch between cbs_ll_heuristic and the associated expander (e.g. txevl_gridmap_expansion_policy). added new expansion policies for the gridmap variants and new compute_h_values functions in cbs_ll_heuristic which use the right map.
also increased the size of ids in search_node objects to 64bit
since the time-expanded search can blow the 32bit limit on nodes

+++ src/search/search_node.h
@@ -222,10 +222,10 @@ class search_node
 		}
 
 	private:
-		uint32_t id_and_status_; // bit 0 is expansion status; 1-31 are id
+		uint64_t id_and_status_; // bit 0 is expansion status; 1-63 are id
 		double f_;
 		double g_;
-        uint32_t parent_id_;
+        uint64_t parent_id_;
 		uint32_t priority_; // expansion priority
 		uint32_t searchid_;
         uint8_t jps_parent_direction_; // hack
